archive/tests/frontend/requirements.txt
# Frontend testing requirements
playwright>=1.40.0
httpx>=0.25.0
google-generativeai>=0.3.1
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0 